Migration from Webpacker to Importmap (#108)
* refactor: Migrated to importmap instead of webpacker

Importmap (https://github.com/rails/importmap-rails) instead of the
retired webpacker

* Added bootstrap in proper importmap manner

* refactor: Removed webpacker gem

Removal of gem introduced a couple a breaking issues that were
also
addressed in this commit:
- Turbolinks needed to replaced by
turbo
- Route to delete needed to follow a new format

* fix: Fixed devise config to work with Turbo

Devise had issues with Turbo and did'nt work out of the box:
https://github.com/heartcombo/devise/issues/5446

Workaround/solution as
shown here:
https://betterprogramming.pub/devise-auth-setup-in-rails-7-44240aaed4be

* fix: Fixed delete format for Program

* Undid changes required for windows local running

* Removed yarn and webpacker steps from Github tests

* Removed webpacker + yarn steps in makefile

* Removed yarn and package.json from Dockerfile

* fix: Fix to destroy tests failing

The addition of the 'Turbo' library changed the way delete alert was shown. For some reason, test was too fast and click action does not work. Adding a 1 second delay before click fixed the test.