Reverted composite primary key of system versioning at rails level
Initially, transaction_end was added as part of a composite key at the rails level using a 3rd party gem. This led to several issues that required workarounds. Since there is no benefit from having rails recognise 'transaction_end' as primary key, having it ignore it is the better approach.